Rename: input 컴포넌트 파일명 소문자로 변경
- Style.ts -> style.ts
- Type.ts -> type.ts
- 누락된 Input 글자 색을 추가
